refactor(core): split table-controls geometry and selection helpers
Extract the two self-contained, internal concerns from the 883-line
table-controls extension into co-located modules:

- table-controls-geometry.ts: boundary and hover detection (pure DOM
  measurement via getBoundingClientRect).
- table-controls-selection.ts: cell selection and table-position lookup.

table-controls.ts keeps the menu definitions, overlay rendering, and the
extension, and imports the extracted helpers. No public symbol moves.
