+++ packages/core/src/extensions/table-controls.ts
@@ -7,6 +7,17 @@ import {
   type VizelTableIconName,
 } from "../icons/types.ts";
 import { VizelTable } from "./table-base";
+import {
+  findColumnBoundary,
+  findHoveredColumn,
+  findHoveredRow,
+  findRowBoundary,
+} from "./table-controls-geometry.ts";
+import {
+  getTablePosition,
+  selectCellInColumn,
+  selectCellInRow,
+} from "./table-controls-selection.ts";
 
 /**
  * Options for table controls UI (column/row insert buttons, row handle menu).
@@ -99,9 +110,6 @@ const CELL_MENU_ITEMS: VizelTableMenuItem[] = [
   { label: "Split cell", command: "splitCell" },
 ];
 
-/** Threshold distance in pixels for detecting boundary hover */
-const BOUNDARY_THRESHOLD_PX = 20;
-
 /** Padding offset constants matching CSS (.vizel-table-controls-wrapper) */
 const PADDING_TOP_PX = 20;
 const PADDING_LEFT_PX = 24;
@@ -116,230 +124,6 @@ const VIZEL_TABLE_MENU_ITEMS: VizelTableMenuItem[] = [
   { divider: true, label: "", command: "" },
   { label: "Delete table", command: "deleteTable", destructive: true },
 ];
-
-/**
- * Find the column boundary index closest to the mouse X position
- */
-function findColumnBoundary(
-  tableElement: HTMLTableElement,
-  mouseX: number,
-  tableRect: DOMRect
-): { index: number; position: number } | null {
-  const firstRow = tableElement.querySelector("tr");
-  if (!firstRow) return null;
-
-  const cells = firstRow.querySelectorAll("td, th");
-  if (cells.length === 0) return null;
-
-  const boundaries: { index: number; position: number }[] = [];
-
-  // Add left boundary (before first column)
-  const firstCell = cells[0];
-  if (!firstCell) return null;
-  const firstCellRect = firstCell.getBoundingClientRect();
-  boundaries.push({ index: 0, position: firstCellRect.left - tableRect.left });
-
-  // Add boundaries between cells and after last cell
-  cells.forEach((cell, i) => {
-    const cellRect = cell.getBoundingClientRect();
-    boundaries.push({ index: i + 1, position: cellRect.right - tableRect.left });
-  });
-
-  // Find closest boundary
-  const relativeX = mouseX - tableRect.left;
-  const firstBoundary = boundaries[0];
-  if (!firstBoundary) return null;
-
-  const { closest, minDistance } = boundaries.reduce(
-    (acc, boundary) => {
-      const distance = Math.abs(relativeX - boundary.position);
-      return distance < acc.minDistance ? { closest: boundary, minDistance: distance } : acc;
-    },
-    { closest: firstBoundary, minDistance: Math.abs(relativeX - firstBoundary.position) }
-  );
-
-  // Only return if within threshold
-  return minDistance <= BOUNDARY_THRESHOLD_PX ? closest : null;
-}
-
-/**
- * Find the row boundary index closest to the mouse Y position
- */
-function findRowBoundary(
-  tableElement: HTMLTableElement,
-  mouseY: number,
-  tableRect: DOMRect
-): { index: number; position: number } | null {
-  const rows = tableElement.querySelectorAll("tr");
-  if (rows.length === 0) return null;
-
-  const boundaries: { index: number; position: number }[] = [];
-
-  // Add top boundary (before first row)
-  const firstRow = rows[0];
-  if (!firstRow) return null;
-  const firstRowRect = firstRow.getBoundingClientRect();
-  boundaries.push({ index: 0, position: firstRowRect.top - tableRect.top });
-
-  // Add boundaries between rows and after last row
-  rows.forEach((row, i) => {
-    const rowRect = row.getBoundingClientRect();
-    boundaries.push({ index: i + 1, position: rowRect.bottom - tableRect.top });
-  });
-
-  // Find closest boundary
-  const relativeY = mouseY - tableRect.top;
-  const firstBoundary = boundaries[0];
-  if (!firstBoundary) return null;
-
-  const { closest, minDistance } = boundaries.reduce(
-    (acc, boundary) => {
-      const distance = Math.abs(relativeY - boundary.position);
-      return distance < acc.minDistance ? { closest: boundary, minDistance: distance } : acc;
-    },
-    { closest: firstBoundary, minDistance: Math.abs(relativeY - firstBoundary.position) }
-  );
-
-  // Only return if within threshold
-  return minDistance <= BOUNDARY_THRESHOLD_PX ? closest : null;
-}
-
-/**
- * Find which row the mouse is hovering over
- */
-function findHoveredRow(
-  tableElement: HTMLTableElement,
-  mouseY: number
-): { index: number; element: HTMLTableRowElement } | null {
-  const rows = tableElement.querySelectorAll("tr");
-
-  for (let i = 0; i < rows.length; i++) {
-    const row = rows[i] as HTMLTableRowElement;
-    const rowRect = row.getBoundingClientRect();
-    if (mouseY >= rowRect.top && mouseY <= rowRect.bottom) {
-      return { index: i, element: row };
-    }
-  }
-
-  return null;
-}
-
-/**
- * Find which column the mouse is hovering over
- */
-function findHoveredColumn(
-  tableElement: HTMLTableElement,
-  mouseX: number,
-  tableRect: DOMRect
-): { index: number; centerX: number } | null {
-  const firstRow = tableElement.querySelector("tr");
-  if (!firstRow) return null;
-
-  const cells = firstRow.querySelectorAll("td, th");
-  if (cells.length === 0) return null;
-
-  for (let i = 0; i < cells.length; i++) {
-    const cell = cells[i];
-    if (!(cell instanceof HTMLElement)) continue;
-    const cellRect = cell.getBoundingClientRect();
-    if (mouseX >= cellRect.left && mouseX <= cellRect.right) {
-      // Return the center X position relative to the table
-      const centerX = cellRect.left + cellRect.width / 2 - tableRect.left;
-      return { index: i, centerX };
-    }
-  }
-
-  return null;
-}
-
-/**
- * Select the first cell in a specific row
- */
-function selectCellInRow(editor: Editor, tablePos: number, rowIndex: number): void {
-  const { state } = editor;
-  const table = state.doc.nodeAt(tablePos);
-  if (!table) return;
-
-  // Sum sizes of preceding rows to locate the requested row's start position.
-  const startOffset = Array.from({ length: rowIndex }, (_, i) => table.child(i).nodeSize).reduce(
-    (sum, size) => sum + size,
-    0
-  );
-  if (rowIndex >= table.childCount) return;
-  const rowStartPos = tablePos + 1 + startOffset;
-  // Skip row node start then enter the first cell.
-  const cellPos = rowStartPos + 1;
-  editor
-    .chain()
-    .focus()
-    .setTextSelection(cellPos + 1)
-    .run();
-}
-
-/**
- * Select a cell in a specific column of the first row
- */
-function selectCellInColumn(editor: Editor, tablePos: number, colIndex: number): void {
-  const { state } = editor;
-  const table = state.doc.nodeAt(tablePos);
-  if (!table || table.childCount === 0) return;
-
-  const firstRow = table.child(0);
-  const targetIndex = colIndex > firstRow.childCount ? firstRow.childCount - 1 : colIndex;
-  if (targetIndex < 0 || targetIndex >= firstRow.childCount) {
-    // If colIndex is 0 (and clamping kept us here), select first cell.
-    if (colIndex === 0 && firstRow.childCount > 0) {
-      const firstCellPos = tablePos + 3;
-      editor.chain().focus().setTextSelection(firstCellPos).run();
-    }
-    return;
-  }
-
-  // Sum sizes of preceding cells to locate the target cell's start position.
-  const cellOffset = Array.from(
-    { length: targetIndex },
-    (_, i) => firstRow.child(i).nodeSize
-  ).reduce((sum, size) => sum + size, 0);
-
-  // Skip table node start and row node start, then accumulate cell sizes.
-  const pos = tablePos + 2 + cellOffset;
-  editor
-    .chain()
-    .focus()
-    .setTextSelection(pos + 1)
-    .run();
-}
-
-/**
- * Find table position from editor selection by walking up the document tree.
- * Used as fallback when getPos() returns undefined.
- */
-function findTablePositionFromSelection(editor: Editor): number | undefined {
-  try {
-    const { selection } = editor.state;
-    const resolved = selection.$from;
-    for (let d = resolved.depth; d > 0; d--) {
-      const node = resolved.node(d);
-      if (node.type.name === "table") {
-        return resolved.before(d);
-      }
-    }
-  } catch {
-    // Ignore errors in fallback
-  }
-  return undefined;
-}
-
-/**
- * Get table position from getPos function or fallback to selection-based lookup.
- */
-function getTablePosition(
-  getPos: (() => number | undefined) | boolean,
-  editor: Editor
-): number | undefined {
-  const fromGetPos = typeof getPos === "function" ? getPos() : undefined;
-  return fromGetPos ?? findTablePositionFromSelection(editor);
-}
 
 /**
  * Set text alignment for all cells in a specific column (Markdown compatible)
